refactor: s/size_t/int/

+++ libxtd/array/array.h
@@ -20,8 +20,8 @@ extern "C"
 #endif                                 /* C++ */
     typedef struct ArrayContainer_t
     {
-        size_t n_items;
-        size_t item_size;
+        int n_items;
+        int item_size;
         char *base;                    /* size == n_items*item_size */
     } ArrayContainer, *ArrayContainerPtr;
 
@@ -35,7 +35,7 @@ extern "C"
      * base --the base address of the array storage
      */
     inline void array_init(ArrayContainerPtr array,
-                           size_t n_items, size_t item_size, void *base)
+                           int n_items, int item_size, void *base)
     {
         array->n_items = n_items;
         array->item_size = item_size;
@@ -52,7 +52,7 @@ extern "C"
      * Returns: (void *)
      * The address of item[offset].
      */
-    inline void *array_item(ArrayContainerPtr array, ssize_t offset)
+    inline void *array_item(ArrayContainerPtr array, int offset)
     {
         return array->base + offset * (long) array->item_size;
     }
