+++ libxtd/array/heap.h
@@ -14,13 +14,13 @@ extern "C"
     typedef struct Heap_t
     {
         ArrayContainer array;
-        size_t n_used;                 /* current size of heap */
+        int n_used;                 /* current size of heap */
         CompareProc cmp;               /* heap item comparison function */
     } Heap, *HeapPtr;
 
     HeapPtr heap_alloc(void);
-    HeapPtr heap_init(HeapPtr heap, CompareProc cmp, size_t n_items,
-                      size_t item_size, void *items);
+    HeapPtr heap_init(HeapPtr heap, CompareProc cmp, int n_items,
+                      int item_size, void *items);
 
     int heap_push(HeapPtr heap, const void *item);
     int heap_pop(HeapPtr heap, void *item);
@@ -35,11 +35,11 @@ extern "C"
     /*
      * low-level heap operations.
      */
-    int heap_ok(void *heap, size_t n_items, size_t item_size,
+    int heap_ok(void *heap, int n_items, int item_size,
                 CompareProc cmp);
-    void heap_sift_up(void *heap, size_t n_items, size_t item_size,
+    void heap_sift_up(void *heap, int n_items, int item_size,
                       CompareProc cmp);
-    void heap_sift_down(void *heap, size_t n_items, size_t item_size,
+    void heap_sift_down(void *heap, int n_items, int item_size,
                         CompareProc cmp);
 #ifdef __cplusplus
 }
