+++ libxtd/array/stack.h
@@ -13,12 +13,12 @@ extern "C"
     typedef struct Stack_t
     {
         ArrayContainer array;
-        size_t top;                    /* current top of stack */
+        int top;                    /* current top of stack */
     } Stack, *StackPtr;
 
     StackPtr stack_alloc(void);
-    StackPtr stack_init(StackPtr stack, size_t n_items,
-                        size_t item_size, void *items);
+    StackPtr stack_init(StackPtr stack, int n_items,
+                        int item_size, void *items);
 
     int stack_push(StackPtr stack, const void *item);
     int stack_pop(StackPtr stack, void *item);
